asio demo updates
